refactor connection pool
